--- sdk/oldconfig/doc.go
@@ -1,7 +1,0 @@
-/*
-Package config contains the configuration definitions and utilities for the SDK.
-
-In addition to defining the structs that act as the configuration schemes, this
-package also implements config versioning, validation, and verification methods.
-*/
-package oldconfig
